refactor(plugins): split upgrade.ts into lane modules
src/core/plugins/upgrade.ts (895 lines) becomes the ~200-line public
entry point (upgradePlugin orchestrator + the local lane) over four
sibling modules, moved byte-for-byte:

- upgrade-gate.ts — UpgradeOptions/UpgradeResult/UpgradeRefusedError,
  the security audit trail, the run() git wrapper, readManifest, and the
  id-stability / signature-policy / permission gates every lane runs
- upgrade-github.ts — origin-pinning + ff-only fast-forward lane and the
  subpath staging-clone lane (force-push defense intact)
- upgrade-artifact.ts — isArtifactInstall, latestPublishedVersion, and
  the Whiskit published-artifact lane (lock/staging/commit unchanged)
- upgrade-check.ts — the read-only --check probe with its single-batched-
  lockfile-write contract (probeOne stays strictly read-only)
- source-tree-sha.ts — computeSourceTreeSha (algorithm unchanged here;
  the hasher consolidation lands separately)

upgrade.ts re-exports the full prior surface (UpgradeRefusedError,
runChecks, UpgradeAvailability, isArtifactInstall, computeSourceTreeSha)
so the host API importers and the '@/core/plugins/upgrade' test mocks
keep working without churn.
